requirements.txt
# Django RAG portal — Azure OpenAI + Azure AI Search
django>=5.0,<6
django-ratelimit>=4.1.0
dj-database-url>=2.2.0
mozilla-django-oidc>=4.0.1
azure-identity>=1.18.0
azure-keyvault-secrets>=4.8.0
openai>=1.40.0
azure-search-documents>=11.6.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
gunicorn>=22.0.0
whitenoise>=6.7.0
pypdf>=4.0.0
python-docx>=1.1.0
psycopg[binary]>=3.2.0
ruff>=0.6.0
mypy>=1.11.0
bandit>=1.7.10
